+++ packages/infra/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @effect-app/infra
 
+## 4.0.0-beta.182
+
+### Patch Changes
+
+- effect-app@4.0.0-beta.182
+
+## 4.0.0-beta.181
+
+### Patch Changes
+
+- Updated dependencies [583393f]
+  - effect-app@4.0.0-beta.181
+
 ## 4.0.0-beta.180
 
 ### Minor Changes